+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -184,7 +184,7 @@ export function withExtendedResolver(
   const defaultResolver = metroResolver.resolve;
   const resolver = isFastResolverEnabled
     ? createFastResolver({
-        preserveSymlinks: config.resolver?.unstable_enableSymlinks ?? true,
+        preserveSymlinks: true,
         blockList: Array.isArray(config.resolver?.blockList)
           ? config.resolver?.blockList
           : [config.resolver?.blockList],