+++ packages/expo-navigation-bar/CHANGELOG.md
@@ -15,6 +15,10 @@
 
 - Removed redundant usage of `EventEmitter` instance. ([#28946](https://github.com/expo/expo/pull/28946) by [@tsapeta](https://github.com/tsapeta))
 
+### ⚠️ Notices
+
+- Added support for React Native 0.76.x. ([#31552](https://github.com/expo/expo/pull/31552) by [@gabrieldonadel](https://github.com/gabrieldonadel))
+
 ## 3.0.7 - 2024-07-03
 
 _This version does not introduce any user-facing changes._